+++ sacrerouge/datasets/nytimes/__init__.py
@@ -1,0 +1,1 @@
+from sacrerouge.datasets.nytimes.subcommand import NYTimesSubcommand