+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Added the New York Times dataset. See [here](doc/datasets/nytimes.md).
+
 ### Changed
 - Added an exception with an error message if PyrEval is used with a single reference summary.